Expose Pipe{Reader|Writer} and their Close[WithError] methods, like io.Pipe
This allows library users to access the Close methods without funny
interface assertions, and replicate functionality similar (but not
identical) to NewReader.  It also mirrors the io.Pipe implementation, and
follows the general best practice of returning concretes instead of
interfaces.

This does change an exposed API, but in such a way that should not break
most existing code, as shown by the unchanged tests.